fix: retry to move telemetry.ts (and save files this time)
