+++ extension/src/PermissionSet/PermissionSetNamePanel.ts
@@ -2,7 +2,7 @@ import * as vscode from "vscode";
 import * as html from "../XliffEditor/HTML";
 import * as SettingsLoader from "../Settings/SettingsLoader";
 import * as PermissionSetFunctions from "./PermissionSetFunctions";
-import * as Telemetry from "../Telemetry";
+import * as Telemetry from "../Telemetry/Telemetry";
 import { XmlPermissionSet } from "./XmlPermissionSet";
 import { logger } from "../Logging/LogHelper";
